Move workflow actions to separate view
