+++ src/views/workflow-actions/workflow-actions-menu/workflow-actions-menu.types.ts
@@ -1,0 +1,8 @@
+import { type DescribeWorkflowResponse } from '@/route-handlers/describe-workflow/describe-workflow.types';
+
+import { type WorkflowAction } from '../workflow-actions.types';
+
+export type Props = {
+  workflow: DescribeWorkflowResponse;
+  onActionSelect: (action: WorkflowAction) => void;
+};
