--- src/views/workflow-page/workflow-page-actions-button/workflow-page-actions-button.types.ts
@@ -1,6 +1,0 @@
-export type Props = {
-  domain: string;
-  cluster: string;
-  workflowId: string;
-  runId: string;
-};
